Expand extra context lines only up to file length
It didn't make sense to try to mark more surrounding lines as edited
than there are lines of text in the file.

Remove the 1000-line hard cap for the number of extra context lines.

Theoretically this could prolong the run time for very long
pathological files whose AST partial reformatting fails to maintain
unchanged. Good luck trying to come up with such a case, though.  If
this should happen, it would most probably reveal a bug in Black
instead.