Allow to give a commitish reference.
This allow to give a commit-ish reference and can thus work across
multiple commits.

I find this convenient as I usually want to run darker after having
committed or on a pull-request with many changes. I'm also thinking that
for example  if the commit-ish is `origin/master`, or in a config file,
some commit at some point in time; eventually all the lines will be
touched and this will become equivalent to isort + black.

I'm opened to better names for the option.