Optionally print the diff in color.
When pygments is installed and stdout is a tty, print the diff in color.
I doubt an option to force color/nocolor is necessary, but this could be
slightly friendlier for people checking diff manually.

We ignore pygments in mypy; as there is not type anotation, the
whitespacing is slightly altered.

Also always print the 10 slowest test with pytest.