Unified text document class
We used to handle text documents either as strings or as lists of
strings, or both, depending on in which format various 3rd party
libraries needed them.

To simplify things, a `TextDocument` class has now been created. It
automatically provides the contents of a document both in a string and
a tuple of strings format. The objects are immutable, and tuples are
used to achieve immutability of the text line collections.