Handle files not encoded with UTF-8
Black can format files encoded with an encoding other than UTF-8, based
on the encoding cookie present in the first two lines. Darker is also
supposed to format those files, but an exception is raised instead.

The exception happens because all the output from git is decoded with
UTF-8, including calls to `git show <revision>:<path>` that returns the
binary content of a file from the history. If a file used an encoding
other than UTF-8, the decoding fails.

`TextDocument.from_bytes()` can now create a `TextDocument` from the raw
binary content of a file, automatically detecting the encoding used.
From now on, this new factory is used when loading a file from disk or
from a revision in the git history.