Create the test repos with --initial-branch=master
git prints a warning when creating a new repository if
`init.defaultBranch` is not configured and `git init` is called without
`--initial-branch`.

Furthermore, several tests expect the initial branch to be `master`.
Those tests will fail if the default configuration for the default
branch ever changes.

Therefore, the repositories created for testing are explicitly
configured to use the initial branch `master`.