Create test files in subdirs
this reduces the number of files for the subprocess test from 1100 to 210. A bit more sane.

Unfortunately we also fall foul of the Windows path length limit here, so we can't have any number of directories.

Also mark the test for this bug to be skipped on non windows systems as it's not an issue there.